Store Complete, New Camera Functions, Teleport Spell
Store is now fully functioning and in the main game, There is also a new
spell that allows the player to teleport to the store, the camera also
now includes Smooth Damping, and has the ability to zoom in and out with
the mouse wheel. Soul counter is also working.